Separate out a document registry
